Assignment 01 refactoring for GitHub classroom complete
Most of the refactoring work for assignment 01 to work with new
GitHub classroom repository work flow is done in this commit.

- Restructured assignment description for new workflow.
- Added a few more tasks/steps for assignment work.
- Added in tests for the new assignment steps.

+++ src/assg01-main.cpp
@@ -1,10 +1,10 @@
 /** @file assg01-main.cpp
  * @brief main/debug executable for Assignment 01, practice on functions.
  *
- * @author Jane Programmer
+ * @author Derek Harter
  * @note   cwid : 123 45 678
  * @note   class: COSC 2336, Summer 2021
- * @note   ide  : VS Code Server 3.9.3 
+ * @note   ide  : VS Code Server 3.9.3
  * @note   assg : Assignment 01
  * @date   June 1, 2021
  *
@@ -34,14 +34,14 @@ int main(int argc, char** argv)
 {
   // uncomment declaration of the x array when you want to compile and call
   // your functions.
-  //int x[] = {3, 5, 7};
+  //double x[] = {-5, -3, -1, 1, 3, 5};
 
-  int mean = 0;
+  double mean = 0;
   // if/when ready to debug calculateMean() uncomment the following and recompile
   //mean = calculateMean(3, x);
   cout << "The mean value is: " << mean << endl;
 
-  int std = 0;
+  double std = 0;
   // if/when ready to debug calculateStandardDeviation() uncomment the following
   // and recompile
   //std = calculateStandardDeviation(1, x);
